requirements.txt
asciichartpy==1.5.25
colorama==0.4.6
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.6
opencv-python==4.11.0.86
pyfiglet==1.0.2
Pygments==2.19.1
python-vlc==3.0.21203
rich==14.0.0
setuptools==80.8.0
termgraph==0.5.3
